+++ app/builders/finance_transactional_report_builder.rb
@@ -1,7 +1,4 @@
-class FinanceTransactionalReportBuilder
-  require 'csv'
-  require 'zip'
-  attr_reader :zipfile_path
+class FinanceTransactionalReportBuilder < ReportBase
 
   CSV_FIELDS = {
     month_year: 'Month-Year',
@@ -25,21 +22,6 @@ class FinanceTransactionalReportBuilder
 
     @csv_file_name = "finance-transactional-#{start_date.values.join('-')}-#{end_date.values.join('-')}.csv"
     @zipfile_path = "tmp/#{@csv_file_name}.zip"
-  end
-
-  def format_dates(date_attribute)
-    DateTime.parse(date_attribute.values.join('/')).utc
-  end
-
-  def to_zip
-    @csv_data = to_csv
-    generate_file
-  end
-
-  def generate_file
-    Zip::File.open(@zipfile_path, Zip::File::CREATE) do |zipfile|
-      zipfile.get_output_stream(@csv_file_name) { |f| f.write @csv_data }
-    end
   end
 
   def to_csv
