+++ app/models/views/reports/raw_data_export.rb
@@ -1,10 +1,7 @@
 # rubocop:disable Metrics/ClassLength
 module Views
   module Reports
-    class RawDataExport
-      require 'csv'
-      require 'zip'
-      attr_reader :zipfile_path
+    class RawDataExport < ReportBase
 
       FIELDS = {
         id: 'id',
@@ -59,21 +56,6 @@ module Views
 
         @csv_file_name = "raw_data-#{start_date.values.join('-')}-#{end_date.values.join('-')}.csv"
         @zipfile_path = "tmp/#{@csv_file_name}.zip"
-      end
-
-      def format_dates(date_attribute)
-        DateTime.parse(date_attribute.values.join('/')).utc
-      end
-
-      def to_zip
-        @csv_data = to_csv
-        generate_file
-      end
-
-      def generate_file
-        Zip::File.open(@zipfile_path, Zip::File::CREATE) do |zipfile|
-          zipfile.get_output_stream(@csv_file_name) { |f| f.write @csv_data }
-        end
       end
 
       def to_csv
